Confirmation when deleting domains and templates

+++ domaintable.inc.php
@@ -135,9 +135,9 @@ function showdomains ($count, $page, $adminlist, $search, $user) {
 			<? domainowners($row->domain_id) ?>
 			</select>
 		</td>
-		<td class="actions">[ <a href="domain-delete.php?id=<?php print $row->domain_id; ?>" onClick="return confirmAction('Delete this domain?')">Delete</a> | <a href="editdomain.php?id=<?php print $row->domain_id; ?>">Edit</a> ]</td>
+		<td class="actions">[ <a href="domain-delete.php?id=<?php print $row->domain_id; ?>" onClick="return confirmAction('Delete domain: <?php print htmlentities($row->domainname); ?>?')">Delete</a> | <a href="editdomain.php?id=<?php print $row->domain_id; ?>">Edit</a> ]</td>
         </tr>
-<?	
+<?
 #          }
 	}
 ?>        </table>
@@ -288,7 +288,7 @@ function showdomain ($domainid, $count, $page, $adminlist, $search) {
 		<td class="content"><?php print $row->content; ?></td>
 		<td class="ttl"><?php print $row->ttl; ?></td>
 		<td class="priority"><?php print $row->prio; ?></td>
-		<td class="actions">[<a href="record-delete.php?type=0&id=<?php print $row->id; ?>">Delete</a> | <a href="editrecord.php?type=0&id=<?php print $row->id; ?>">Edit</a>]</td>
+		<td class="actions">[<a href="record-delete.php?type=0&id=<?php print $row->id; ?>" onClick="return confirmAction('Delete this record?')">Delete</a> | <a href="editrecord.php?type=0&id=<?php print $row->id; ?>">Edit</a>]</td>
         </tr>
 <?
 	}
